Replaced Octokit GH Client with the new GitHub client for branches sync
