+++ package.json
@@ -48,6 +48,7 @@
     "@atlaskit/reduced-ui-pack": "^8.19.0",
     "@atlassian/aui": "^9.4.0",
     "@octokit/auth-app": "^2.6.0",
+    "@octokit/graphql-schema": "^10.73.0",
     "@sentry/node": "^6.8.0",
     "@types/bunyan": "^1.8.6",
     "@types/bunyan-format": "^0.2.4",
