+++ src/config/feature-flags.ts
@@ -23,7 +23,8 @@ export enum BooleanFlags {
 	SUPPORT_BRANCH_AND_MERGE_WORKFLOWS_FOR_BUILDS = "support-branch-and-merge-workflows-for-builds",
 	USE_SQS_FOR_BRANCH = "use-sqs-for-branch",
 	ASSOCIATE_PR_TO_ISSUES_IN_BODY = "associate-pr-to-issues-in-body",
-	VERBOSE_LOGGING = "verbose-logging"
+	VERBOSE_LOGGING = "verbose-logging",
+	USE_NEW_GITHUB_CLIENT_FOR_BRANCHES = "use-new-github-client-for-branches"
 }
 
 export enum StringFlags {
