+++ src/sync/branches.ts
@@ -1,21 +1,35 @@
 import transformBranches from "./transforms/branch";
-import { getBranches as getBranchesQuery } from "./queries";
 import { GitHubAPI } from "probot";
 import { Repository } from "../models/subscription";
+import { Repository as OctokitRepository} from "@octokit/graphql-schema";
 import GitHubClient from "../github/client/github-client";
+import { booleanFlag, BooleanFlags } from "../config/feature-flags";
 import { LoggerWithTarget } from "probot/lib/wrap-logger";
+import {getBranches as getBranchesQuery} from "../github/client/github-queries";
 
 // TODO: better typings
-export default async (logger: LoggerWithTarget, github: GitHubAPI, _newGithub: GitHubClient, _jiraHost: string, repository:Repository, cursor?:string | number, perPage?:number) => {
+export default async (logger: LoggerWithTarget, github: GitHubAPI, newGithub: GitHubClient, jiraHost: string, repository:Repository, cursor?:string | number, perPage?:number) => {
 	// TODO: fix typings for graphql
 	logger.info("Syncing branches: started");
-	const { edges } = ((await github.graphql(getBranchesQuery, {
-		owner: repository.owner.login,
-		repo: repository.name,
-		per_page: perPage,
-		cursor
-		// eslint-disable-next-line @typescript-eslint/no-explicit-any
-	})) as any).repository.refs;
+
+	const useNewGHClient = await booleanFlag(BooleanFlags.USE_NEW_GITHUB_CLIENT_FOR_BRANCHES, false, jiraHost);
+
+	let edges;
+	if(useNewGHClient) {
+		const result = await newGithub.getBranchesPage(repository.owner.login, repository.name, perPage, cursor as string)
+		edges = result?.refs?.edges || [];
+	} else {
+
+		const results = ((await github.graphql(getBranchesQuery, {
+			owner: repository.owner.login,
+			repo: repository.name,
+			per_page: perPage,
+			cursor
+			// eslint-disable-next-line @typescript-eslint/no-explicit-any
+		})) as { repository: OctokitRepository});
+
+		edges = results.repository?.refs?.edges || [];
+	}
 
 	const branches = edges.map(({ node: item }) => item);
