+++ src/sync/commits.ts
@@ -1,9 +1,9 @@
 import transformCommit from "../transforms/commit";
-import { getCommits as getCommitsQuery, getDefaultRef } from "./queries";
 import { GitHubAPI } from "probot";
 import { Repository } from "../models/subscription";
 import GitHubClient from "../github/client/github-client";
 import { LoggerWithTarget } from "probot/lib/wrap-logger";
+import {getCommits as getCommitsQuery, getDefaultRef} from "../github/client/github-queries";
 
 // TODO: better typings
 export default async (logger: LoggerWithTarget, github: GitHubAPI, _newGithub: GitHubClient, _jiraHost: string, repository: Repository, cursor?: string | number, perPage?: number) => {
